Port contrib to Scala.js
