+++ scalameta/contrib/jvm/src/test/scala/scala/meta/contrib/ContribSuite.scala
