+++ scalameta/contrib/jvm/src/test/scala/scala/meta/contrib/TreeOpsProps.scala
@@ -1,0 +1,15 @@
+package scala.meta
+package contrib
+
+import scala.meta.testkit.SyntaxAnalysis
+import org.scalatest.FunSuite
+
+class TreeOpsProps extends FunSuite {
+  test("find uses Structural equality") {
+    val errors = SyntaxAnalysis.onParsed[Tree](ContribSuite.corpus) { ast =>
+      val names = ast.collect { case t: Defn.Val => t.copy(pats = t.pats) }
+      names.filterNot(name => ast.contains[Structurally](name))
+    }
+    assert(errors.isEmpty)
+  }
+}
