+++ scalameta/contrib/shared/src/main/scala/scala/meta/contrib/DocToken.scala
@@ -2,6 +2,10 @@ package scala.meta.contrib
 
 import scala.util.matching.Regex
 
+import fastparse.all._
+import fastparse.core.Parsed
+import org.scalameta.logger
+
 /**
   * Represents a scaladoc line.
   */
@@ -20,13 +24,16 @@ case class DocToken(kind: DocToken.Kind, name: Option[String], body: Option[Stri
     */
   def references: Seq[DocToken.Reference] =
     body
-      .map(
-        DocToken.referenceRegex
-          .findAllMatchIn(_)
-          .map(_.group(1))
-          .map(DocToken.Reference)
-          .toList
-      )
+      .map { b =>
+        def parseBodyFrom(idx: Int): Seq[DocToken.Reference] = {
+          DocToken.referenceParser.parse(b, idx) match {
+            case Parsed.Success(value, index) =>
+              Seq(DocToken.Reference(value)) ++ parseBodyFrom(index)
+            case _ => Seq[DocToken.Reference]()
+          }
+        }
+        parseBodyFrom(0)
+      }
       .getOrElse(Nil)
 }
 
@@ -41,7 +48,14 @@ object DocToken {
   /**
     * Parser that for obtaining a class reference from an scaladoc body.
     */
-  private val referenceRegex: Regex = "\\[\\[([^]]+)\\]\\]".r
+  private val referenceParser: Parser[String] = P(
+    // Removes the elements previous to the references.
+    ((AnyChar ~ !"[[").rep ~ AnyChar).?
+    // Retrieves the element within a
+      ~ "[[" ~
+      ((AnyChar ~ !"]]").rep ~ AnyChar).!
+      ~ "]]"
+  )
 
   /**
     * Represents an documentation code reference.
